Spell gammaMinimumX's literals as Int64 for 32-bit platforms
A bare integer literal passed to the generic
Element.init<Source: BinaryInteger> defaults to Int, and on watchOS
(armv7k/arm64_32) and wasm32 Int is 32 bits, so both 16-digit literals
overflowed at compile time -- the two Swift Package Index build
failures.  `as Int64` pins the literal's type on every platform; the
computed value is bit-identical where it already worked.

+++ Sources/Interval/IntervalMath.swift
@@ -174,7 +174,9 @@ extension Interval {
     /// curve's flat bottom, so the critical value it yields is second-order
     /// accurate.
     private static var gammaMinimumX:Element {
-        return Element(1461632144968362) / Element(1000000000000000)
+        // spelled as Int64: a bare literal is an Int, which these overflow on
+        // 32-bit platforms (watchOS armv7k/arm64_32, wasm32)
+        return Element(1461632144968362 as Int64) / Element(1000000000000000 as Int64)
     }
     // gamma - on x > 0, unimodal with its minimum at gammaMinimumX; for x <= 0,
     // computed through the reflection formula Γ(x) = π / (sin(πx)·Γ(1-x)),
